tools/opai-dam/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
pyjwt>=2.8.0
sse-starlette>=1.6.0
anthropic>=0.39.0
